+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/UserServiceApplication.java
@@ -1,4 +1,4 @@
-package n11bootcamp_project_backend.api_gateway;
+package n11bootcamp_project_backend.user_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class ApiGatewayApplication {
+public class UserServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ApiGatewayApplication.class, args);
+		SpringApplication.run(UserServiceApplication.class, args);
 	}
 
 }