+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/dto/request/LoginRequest.java
@@ -1,0 +1,6 @@
+package n11bootcamp_project_backend.user_service.dto.request;
+
+public record LoginRequest(
+        String email,
+        String password
+) {}