+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/dto/request/RegisterRequest.java
@@ -1,0 +1,8 @@
+package n11bootcamp_project_backend.user_service.dto.request;
+
+public record RegisterRequest(
+        String firstName,
+        String lastName,
+        String email,
+        String password
+) {}