+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/dto/response/TokenResponse.java
@@ -1,0 +1,7 @@
+package n11bootcamp_project_backend.user_service.dto.response;
+
+public record TokenResponse(
+        String accessToken,
+        String refreshToken,
+        long expiresIn
+) {}