+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/enums/Role.java
@@ -1,0 +1,6 @@
+package n11bootcamp_project_backend.user_service.enums;
+
+public enum Role {
+    CUSTOMER,
+    ADMIN
+}