+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package n11bootcamp_project_backend.user_service.repository;
+
+import n11bootcamp_project_backend.user_service.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+
+    Boolean existsByEmail(String email);
+
+    Optional<User> findByEmail(String email);
+}