+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/services/UserService.java
@@ -1,0 +1,4 @@
+package n11bootcamp_project_backend.user_service.services;
+
+public interface UserService {
+}