+++ user-service/src/main/java/n11bootcamp_project_backend/user_service/services/impl/UserServiceImpl.java
@@ -1,0 +1,8 @@
+package n11bootcamp_project_backend.user_service.services.impl;
+
+import n11bootcamp_project_backend.user_service.services.UserService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserServiceImpl implements UserService {
+}