+++ user-service/src/test/java/n11bootcamp_project_backend/user_service/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package n11bootcamp_project_backend.user_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}